requirements.txt
openai>=1.0.0
azure-identity>=1.17.0
agent-framework
